requirements.txt
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
plotly==5.3.1
psycopg2-binary==2.9.1
pymongo==3.12.0
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
sklearn==0.0
